upstream release

Merge subtree 'emacspeak'

+++ emacspeak/PKGBUILD
@@ -1,32 +1,31 @@
-# Maintainer: Alexander Epaneshnikov <alex19ep@archlinux.org>
+# Maintainer: Storm Dragon <storm_dragon@stormux.org>
+# Contributor: Alexander Epaneshnikov <alex19ep@archlinux.org>
 # Contributor: Steve Holmes <steve.holmes88@gmail.com>
 # Contributor: Chris Brannon <cmbrannon79@gmail.com>
 
 pkgname=emacspeak
-pkgver=55.0
+pkgver=60.0
 pkgrel=1
 pkgdesc="Emacs extension that provides spoken output"
 arch=('x86_64' 'aarch64')
 url="http://emacspeak.sf.net/"
 license=('GPL' 'LGPL' 'APACHE')
-depends=('emacs' 'tcl' 'tclx' 'espeak-ng')
+depends=('emacs' 'tcl' 'tclx' 'espeak-ng' 'sox')
 optdepends=('python: for Stream The World'
             'perl')
 source=("https://github.com/tvraman/emacspeak/releases/download/${pkgver}/emacspeak-${pkgver}.tar.bz2"
-        "${pkgname}-54.0-directorys.patch"
-        "${pkgname}-54.0-flags.patch"
+        "${pkgname}-60.0-directorys.patch"
+        "${pkgname}-60.0-flags.patch"
         "${pkgname}.sh")
-sha512sums=('2b9424b9d85da8d18803d45403a6e26262e2ed6a51e8868aed2c906c9bd50b9aad797cab8575dba229c3d8f953979d4cf3ed5a0ee16ea8da3e0593e7b040a84f'
-            '870aeac1b91e76ac141983bedd6ed328475a8bcd268d5ec5f48a04b5c285ff29ac6d00681a00350212aa07bba54127a0aa91524d2197e785f12dea059e93ca52'
-            'b287dbe2cc07c29c6d14b79d2984106d5dfc6bebc7636ed0bb937a9c02db926c6ed1cb0b1aa3026ca6b7700514ae5f94c2c39907b42e0594eff82c4ad20ac02f'
-            '3624a1205e70cceb43953d39925e1c9a351635a909100a050b036640df65cc104546939eb3be017b116dbc4f910a018b6a5e2eb39c66c9e01e9c91dde6fdf5d9')
+sha512sums=('f48040f2dfaf7cb67d779d219c59477aaddb9c8b400666e5e9d9850e75575d5276f5b596ab1e8976057e5078d914a610cd537068ccc748e6c45aef64b02eaba9'
+            'b298f72d42eb4334054e8580dc488ac78f2a96f4728e1b6f646c6371a0846ca19a0f607fbaf73d37ed6502bf76d34a820f7e37fbd1047971514a3976dc88daba'
+            '83343badf4adc9758d3d2d63303e64ebb054d461744d8d6603bcc55eaa040f64d9311d303a70cca1483fed38b9e1efc473d9bf4ddfaf62f9b2761b3e6b1a7076'
+            '69456a0dd45b09469a6d345e5528b3eadef5fc4a379425153aa51a597d9faceef85036ca195eef86ae4c08485e38aeff3b64d88662fad4f5c4999117798537ec')
 
 prepare() {
   cd "$pkgname-$pkgver"
-  patch --forward --strip=1 --input="${srcdir}/${pkgname}-54.0-directorys.patch"
-  patch --forward --strip=1 --input="${srcdir}/${pkgname}-54.0-flags.patch"
-  # prebuilt binary
-  rm -v etc/pickup-c/pickup
+  patch --forward --strip=1 --input="${srcdir}/${pkgname}-60.0-directorys.patch"
+  patch --forward --strip=1 --input="${srcdir}/${pkgname}-60.0-flags.patch"
 }
 
 build() {
@@ -44,28 +43,31 @@ package() {
   # clean unneeded files
   find . -name Makefile -delete
   rm -vr info/auto
-  rm -vr sounds/3d/src
+  [ -d sounds/3d/src ] && rm -vr sounds/3d/src || true
 
   install -vDm 644 etc/forms/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/etc/forms"
   install -vDm 644 etc/pickup-c/pickup -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/etc/pickup-c"
   install -vDm 644 etc/tables/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/etc/tables"
+  install -vDm 644 etc/ai/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/etc/ai"
+  find etc/pipewire -type f -exec install -vDm 644 {} "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/{}" \;
+  install -vDm 644 etc/pulse/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/etc/pulse"
   rm -vr etc/forms/
   rm -vr etc/pickup-c/
   rm -vr etc/tables/
-  install -vDm 644 etc/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/etc"
+  rm -vr etc/ai/
+  rm -vr etc/pipewire/
+  rm -vr etc/pulse/
+  find etc/ -maxdepth 1 -type f -exec install -vDm 644 {} -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/etc" \;
   install -vDm 644 info/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/info"
   install -vDm 644 lisp/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/lisp"
   install -vDm 644 xsl/* -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/xsl"
   install -vDm 644 README -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}"
 
-  install -vDm 644 sounds/3d/* -t "${pkgdir}/usr/share/sounds/${pkgname}/3d"
-  install -vDm 644 sounds/classic/* -t "${pkgdir}/usr/share/sounds/${pkgname}/classic"
-  install -vDm 644 sounds/pan-chimes/* -t "${pkgdir}/usr/share/sounds/${pkgname}/pan-chimes"
-  install -vDm 644 sounds/prompts/* -t "${pkgdir}/usr/share/sounds/${pkgname}/prompts"
-  install -vDm 644 sounds/system/* -t "${pkgdir}/usr/share/sounds/${pkgname}/system"
-  install -vDm 644 sounds/{emacspeak.mp3,highbells.au} -t "${pkgdir}/usr/share/sounds/${pkgname}"
+  [ -d sounds/prompts ] && install -vDm 644 sounds/prompts/* -t "${pkgdir}/usr/share/sounds/${pkgname}/prompts"
+  find sounds/ -maxdepth 1 -type f -exec install -vDm 644 {} -t "${pkgdir}/usr/share/sounds/${pkgname}" \;
 
   # speech server
+  install -vDm 644 servers/.servers -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/servers"
   install -vDm 755 servers/tts-lib.tcl -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/servers"
   install -vDm 755 servers/{espeak,log-espeak} -t "${pkgdir}/usr/share/emacs/site-lisp/${pkgname}/servers"
   install -vDm 755 servers/native-espeak/tclespeak.so -t "${pkgdir}/usr/lib/${pkgname}"
